+++ tests/TestCase/Event/EventTest.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
  */
 namespace Tools\Test\Event;
 
-use RuntimeException;
 use Tools\Entity;
 use Tools\Event\Event;
+use Tools\Exception\KeyNotExistsException;
 use Tools\TestSuite\TestCase;
 
 /**
@@ -37,7 +37,7 @@ class EventTest extends TestCase
     {
         parent::setUp();
 
-        $this->Event = new Event('myEvent', ['arg1', 'arg2']);
+        $this->Event = $this->Event ?: new Event('myEvent', ['arg1', 'arg2']);
     }
 
     /**
@@ -58,7 +58,7 @@ class EventTest extends TestCase
         $this->assertSame('arg1', $this->Event->getArg(0));
         $this->assertSame('arg2', $this->Event->getArg(1));
 
-        $this->expectException(RuntimeException::class);
+        $this->expectException(KeyNotExistsException::class);
         $this->expectExceptionMessage('Argument with index `2` does not exist');
         $this->assertSame('arg2', $this->Event->getArg(2));
     }